% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{co_occurrence}
\alias{co_occurrence}
\title{Calculate Normalized Co-occurrence Matrix}
\usage{
co_occurrence(ssp)
}
\arguments{
\item{ssp}{solution space: a list  containing 3 items:
\itemize{
\item \code{M}: A matrix where each column represents a unique community detection solution. Rows are ordered as the nodes in the input graph. Columns are ordered by decreasing probability (median of confidence intervals).
\item \code{data}: A dataframe summarizing the features of each solution, including confidence intervals, similarity scores, and grouping based on confidence levels.
\item \code{simil}: A similarity matrix (using ARI) between the different solutions found.
}}
}
\value{
A symmetric matrix \code{D} where each entry represents the weighted co-occurrence count of node pairs across all trials. The matrix shows how often nodes were grouped together in the same community, scaled by the weights in \code{alpha}.
}
\description{
The \code{co_occurrence} function computes a normalized co-occurrence matrix from a set of community detection solutions. It measures how frequently nodes appear together in the same community across multiple trials and weights the co-occurrence by a specified importance factor \code{alpha} for each trial.
}
\details{
The function uses the results of multiple community detection trials stored in \code{M}. For each trial, nodes that belong to the same community are identified, and their co-occurrence count is incremented by the corresponding value from \code{alpha}. The co-occurrence matrix is symmetric, reflecting the fact that if node A co-occurs with node B, then node B also co-occurs with node A.
}
\examples{
D <- co_occurrence(ssp))
print(co_matrix)

}
