% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{solutions_space}
\alias{solutions_space}
\title{Explore the Solution Space of Community Detection Algorithms}
\usage{
solutions_space(
  g,
  n_trials = 100,
  tau = 0.95,
  met = "IM",
  shuffle = TRUE,
  comp_method = "ami",
  confidence = 0.95,
  resolution = 1,
  IM.nb.trials = 10,
  WT.steps = 3,
  verbose = TRUE
)
}
\arguments{
\item{g}{An iGraph object representing the network on which community detection will be performed.}

\item{n_trials}{An integer specifying the number of trials to run for community detection.
#' @param tau: A value between 0.0 and 1.0 specifying the threshold of stability of the solution space. When p_stable >= tau, the solution spaces is considered stable, i.e. the probability of a nwe solution to be found at the next trial is below tau. and thethe explorarion is terminated. Default value is 0.95.}

\item{met}{A string specifying the community detection method to use. Available options include:
\itemize{
\item \code{"IM"}: Infomap
\item \code{"WT"}: Walktrap
\item \code{"LV"}: Louvain
\item \code{"LD"}: Leiden
\item \code{"LP"}: Label Propagation
\item \code{"EV"}: Leading Eigenvector
\item \code{"EB"}: Edge Betweenness
}}

\item{shuffle}{A logical value indicating whether to shuffle the network (i.e. permute the order of nodes and edges) order before running the community detection algorithm. Default is \code{TRUE}.}

\item{comp_method}{A string specifying the method for comparing solutions. Available options:
\itemize{
\item \code{"ami"}: Adjusted Mutual Information (default)
\item \code{"ari"}: Adjusted Rand Index
}}

\item{confidence}{A numeric value representing the confidence level for estimating the confidence intervals of the solutions. Default is 0.95.}

\item{resolution}{A numeric value for the resolution parameter used by some community detection algorithms, such as Louvain and Leiden. Default is 1.0.}

\item{IM.nb.trials}{An integer specifying the number of trials for the Infomap algorithm. Default is 10.}

\item{WT.steps}{An integer specifying the number of steps for the Walktrap algorithm. Default is 3.}
}
\value{
A list-like object containing 3 items:
\itemize{
\item \code{M}: A matrix where each column represents a unique community detection solution. Rows are ordered as the nodes in the input graph. Columns are ordered by decreasing probability (median of confidence intervals).
\item \code{data}: A dataframe summarizing the features of each solution, including confidence intervals, similarity scores, and grouping based on confidence levels.
\item \code{simil}: A similarity matrix (using ARI) between the different solutions found.
}
}
\description{
The \code{solutions_space} function generates the solution space of a given community detection algorithm, i.e. the set of all independent solutions identified after a number of independent trials. By running multiple trials of a selected community detection method, it identifies one or more unique solutions, and calculates similarity between them. The function also computes the probability associated with each unique solution, using a Bayesian approach (to estimate the confidence of each solution and organizes the results based on confidence intervals)Beta Binomial model).
}
\details{
The function performs multiple trials of the chosen community detection method and compares the solutions using the selected comparison method (AMI or ARI). It checks if new solutions are unique by comparing them to previously found solutions. Unique solutions are added to the result set, and a Bayesian updating procedure is applied to compute confidence intervals for the posterior distribution of each solution. A similarity matrix between solutions is also computed using ARI.

The function supports several popular community detection algorithms, and it allows for shuffling the node order between trials to introduce randomness and explore different solutions.
}
\examples{
# Run 10 trials of the Louvain method and explore the solution space
solution_space <- solutions_space(g, n_trials = 10, tau = 0.9, met = 'LV', comp_method = 'ami')
print(solution_space$data)

}
