% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{empirical_mu}
\alias{empirical_mu}
\title{Calculate the Empirical Mixing Parameter (Mu)}
\usage{
empirical_mu(g, community_labels)
}
\arguments{
\item{g}{An iGraph object representing the network. The edges of the graph should ideally be weighted, but if not, all edge weights are set to 1.0 by default.}

\item{community_labels}{A vector of community labels, ordered according to the vertices in \code{g}. These labels define the community membership of each node in the graph.}
}
\value{
A numeric value representing the empirical mixing parameter (mu), which is the ratio of inter-community edges to the total number of edges.
}
\description{
The \code{empirical_mu} function calculates the mixing parameter "mu" for a given network, which represents the proportion of edges that exist between different communities. This parameter is often used in community detection algorithms to quantify how mixed or modular a network is.
}
\details{
The function works by first ensuring that all edges have weights (defaulting to 1.0 if no weights are provided). It then calculates the number of edges that connect nodes from different communities (inter-community edges) and computes the proportion of such edges relative to the total number of edges in the graph.
}
\examples{
# Create a simple graph and calculate the mixing parameter
g <- make_ring_of_cliques(3, 5)
community_labels <- V(g)$community
mu <- empirical_mu(g, community_labels)
print(mu)

}
