% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{layout_distance_comm}
\alias{layout_distance_comm}
\title{The function layout_distance_gamma calculates a distance between pairs of nodes, that enhances the visualization,
grouping nodes of the same community, and spreading out different communities.}
\usage{
layout_distance_comm(g, membership, eps = 0.02)
}
\arguments{
\item{g:}{The network to be analyzed. It should be an iGraph object with weights E(g)$weight (if the network is unweighted, set E(g)$weight to 1.0)}

\item{membership:}{The membership vector, ordered as V(g).}

\item{eps:}{the distance between community members. typical value in the range 0.1 to 0.2.}
}
\value{
An array of distances, ordered as E(g).
}
\description{
The function layout_distance_gamma calculates a distance between pairs of nodes, that enhances the visualization,
grouping nodes of the same community, and spreading out different communities.
}
