% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ring_of_cliques.R
\name{make_ring_of_cliques}
\alias{make_ring_of_cliques}
\title{Ring of Cliques (RC, RC_B, RC_C, RC_BC)
\itemize{
\item RC:    ring edges from c(k,1) to c(k+1,2)
\item RC_B:  RC + bridge nodes: c(k,2) -- b_k -- c(k+1,3)
\item RC_C:  RC + center connected to c(k,1)
\item RC_BC: RC + bridges + center
}}
\usage{
make_ring_of_cliques(
  n_cliques,
  clique_size,
  variant = c("RC", "RC+C", "RC+B", "RC+BC", "RC_C", "RC_B", "RC_BC"),
  set_names = TRUE
)
}
\arguments{
\item{n_cliques}{integer >= 2}

\item{clique_size}{integer >= 3 (needs nodes 1,2,3)}

\item{variant}{one of "RC","RC_B","RC_C","RC_BC"}

\item{set_names}{logical, add readable names}
}
\value{
igraph
}
\description{
Ring of Cliques (RC, RC_B, RC_C, RC_BC)
\itemize{
\item RC:    ring edges from c(k,1) to c(k+1,2)
\item RC_B:  RC + bridge nodes: c(k,2) -- b_k -- c(k+1,3)
\item RC_C:  RC + center connected to c(k,1)
\item RC_BC: RC + bridges + center
}
}
