% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_sol_space}
\alias{plot_sol_space}
\title{The function plot_solution_space produces plots of the solution space}
\usage{
plot_sol_space(sol_space)
}
\arguments{
\item{sol_space:}{a dataframe produced by explore_solution_space function}
}
\value{
A list of 2 plots
}
\description{
The function plot_solution_space produces plots of the solution space
}
