% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_solutions}
\alias{plot_solutions}
\title{Plot Multiple Solutions from a Solution Space}
\usage{
plot_solutions(
  g,
  ssp,
  device = "screen",
  filename = NULL,
  width = 1600,
  height = 1600,
  res = 300
)
}
\arguments{
\item{g}{An \code{igraph} object representing the network graph.}

\item{device}{A string specifying the output device. Possible values are "screen" (default) and "png".}

\item{filename}{A string specifying the file name for saving the plot if \code{device = "png"}. Default is \code{NULL}.}

\item{width}{Width of the PNG image in pixels. Default is 1600.}

\item{height}{Height of the PNG image in pixels. Default is 1600.}

\item{res}{Resolution of the PNG image in DPI. Default is 300.}

\item{sol_space}{A list containing the solution space. It should include:
\describe{
\item{data}{A matrix with rows corresponding to different solutions.}
\item{M}{A matrix where each column corresponds to the membership of nodes for a given solution.}
}}
}
\value{
NULL. This function is called for creating plots.
}
\description{
The function plot_all_solutions plots all the the solutions from a given solution space on a graph using a grid layout.
Each plot highlights a different solution with communities highlighted in red.
}
\examples{
# Example usage (assume `g` is an igraph object and `sol_space` is defined):
# plot_solutions(g, sol_space)

}
